Remove duplicate code in new proposal components.
