+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalSingle/common/components/NewProposal.tsx
@@ -1,11 +1,15 @@
-import { FlagOutlined, Timelapse } from '@mui/icons-material'
+import { BookOutlined, FlagOutlined, Timelapse } from '@mui/icons-material'
 import { useCallback, useState } from 'react'
+import { useFormContext } from 'react-hook-form'
 import toast from 'react-hot-toast'
 import { useTranslation } from 'react-i18next'
-import { useRecoilCallback, useRecoilValue } from 'recoil'
+import { useRecoilCallback, useRecoilValueLoadable } from 'recoil'
 
 import { DaoCoreV2Selectors, blocksPerYearSelector } from '@dao-dao/state'
 import {
+  NewProposalTitleDescriptionHeader,
+  NewProposal as StatelessNewProposal,
+  NewProposalProps as StatelessNewProposalProps,
   useCachedLoadable,
   useChain,
   useDaoInfoContext,
@@ -15,25 +19,25 @@ import {
   IProposalModuleAdapterCommonOptions,
 } from '@dao-dao/types'
 import {
+  convertActionsToMessages,
   convertExpirationToDate,
   dateToWdhms,
   processError,
 } from '@dao-dao/utils'
 
-import { useLoadedActionsAndCategories } from '../../../../../../actions'
-import { EntityDisplay } from '../../../../../../components/EntityDisplay'
-import { SuspenseLoader } from '../../../../../../components/SuspenseLoader'
-import { useMembership, useWallet } from '../../../../../../hooks'
-import { proposalSelector } from '../../../contracts/DaoProposalMultiple.recoil'
-import { makeGetProposalInfo } from '../../../functions'
+import { useLoadedActionsAndCategories } from '../../../../../actions'
+import { useMembership, useWallet } from '../../../../../hooks'
+import { proposalSelector } from '../../contracts/DaoProposalSingle.common.recoil'
+import { makeGetProposalInfo } from '../../functions'
 import {
   NewProposalData,
   NewProposalForm,
   SimulateProposal,
   UsePublishProposal,
-} from '../../../types'
-import { useProcessQ } from '../../hooks'
-import { NewProposal as StatelessNewProposal } from './NewProposal'
+} from '../../types'
+import { useProcessTQ } from '../hooks'
+import { NewProposalMain } from './NewProposalMain'
+import { NewProposalPreview } from './NewProposalPreview'
 
 export type NewProposalProps = BaseNewProposalProps<NewProposalForm> & {
   options: IProposalModuleAdapterCommonOptions
@@ -55,9 +59,11 @@ export const NewProposal = ({
     isActive,
     activeThreshold,
   } = useDaoInfoContext()
-  const { isWalletConnected, getStargateClient } = useWallet()
+  const { isWalletConnecting, isWalletConnected, getStargateClient } =
+    useWallet()
 
-  const { loadedActions, categories } = useLoadedActionsAndCategories()
+  const { watch } = useFormContext<NewProposalForm>()
+  const proposalTitle = watch('title')
 
   const { isMember = false, loading: membershipLoading } = useMembership({
     coreAddress,
@@ -79,9 +85,9 @@ export const NewProposal = ({
     pauseInfo.state === 'hasValue' &&
     ('paused' in pauseInfo.contents || 'Paused' in pauseInfo.contents)
 
-  const processQ = useProcessQ()
+  const processTQ = useProcessTQ()
 
-  const blocksPerYear = useRecoilValue(
+  const blocksPerYearLoadable = useRecoilValueLoadable(
     blocksPerYearSelector({
       chainId,
     })
@@ -116,6 +122,12 @@ export const NewProposal = ({
           return
         }
 
+        if (blocksPerYearLoadable.state !== 'hasValue') {
+          toast.error(t('error.loadingData'))
+          return
+        }
+        const blocksPerYear = blocksPerYearLoadable.contents
+
         setLoading(true)
         try {
           const { proposalNumber, proposalId } = await publishProposal(
@@ -127,6 +139,7 @@ export const NewProposal = ({
             }
           )
 
+          // Get proposal info to display card.
           const proposalInfo = await makeGetProposalInfo({
             ...options,
             proposalNumber,
@@ -154,7 +167,7 @@ export const NewProposal = ({
             )
           ).proposal
 
-          const { quorum } = processQ(proposal.voting_strategy)
+          const { threshold, quorum } = processTQ(proposal.threshold)
 
           onCreateSuccess(
             proposalInfo
@@ -164,9 +177,17 @@ export const NewProposal = ({
                   description: newProposalData.description,
                   info: [
                     {
-                      Icon: FlagOutlined,
-                      label: `${t('title.quorum')}: ${quorum.display}`,
+                      Icon: BookOutlined,
+                      label: `${t('title.threshold')}: ${threshold.display}`,
                     },
+                    ...(quorum
+                      ? [
+                          {
+                            Icon: FlagOutlined,
+                            label: `${t('title.quorum')}: ${quorum.display}`,
+                          },
+                        ]
+                      : []),
                     ...(expirationDate
                       ? [
                           {
@@ -205,30 +226,44 @@ export const NewProposal = ({
       },
     [
       isWalletConnected,
-      t,
       publishProposal,
       options,
-      blocksPerYear,
+      blocksPerYearLoadable,
       getStargateClient,
       chainId,
-      processQ,
+      processTQ,
       onCreateSuccess,
+      t,
       daoName,
       coreAddress,
       daoImageUrl,
     ]
   )
 
+  const { loadedActions } = useLoadedActionsAndCategories()
+
+  const getProposalDataFromFormData: StatelessNewProposalProps<
+    NewProposalForm,
+    NewProposalData
+  >['getProposalDataFromFormData'] = ({ title, description, actionData }) => ({
+    title,
+    description,
+    msgs: convertActionsToMessages(loadedActions, actionData),
+  })
+
   return (
-    <StatelessNewProposal
-      EntityDisplay={EntityDisplay}
-      SuspenseLoader={SuspenseLoader}
+    <StatelessNewProposal<NewProposalForm, NewProposalData>
       activeThreshold={activeThreshold}
       anyoneCanPropose={anyoneCanPropose}
-      categories={categories}
       connected={isWalletConnected}
+      content={{
+        Header: NewProposalTitleDescriptionHeader,
+        Main: NewProposalMain,
+        Preview: NewProposalPreview,
+      }}
       createProposal={createProposal}
       depositUnsatisfied={depositUnsatisfied}
+      getProposalDataFromFormData={getProposalDataFromFormData}
       isActive={isActive}
       isMember={
         membershipLoading
@@ -236,8 +271,9 @@ export const NewProposal = ({
           : { loading: false, data: isMember }
       }
       isPaused={isPaused}
-      loadedActions={loadedActions}
+      isWalletConnecting={isWalletConnecting}
       loading={loading || simulating}
+      proposalTitle={proposalTitle}
       simulateProposal={simulateProposal}
       simulationBypassExpiration={simulationBypassExpiration}
       {...props}
