+++ packages/stateless/components/proposal/index.ts
@@ -3,6 +3,8 @@ export * from './GovProposalCreatedModal'
 export * from './GovProposalStatus'
 export * from './GovProposalVoteDisplay'
 export * from './GovProposalWalletVote'
+export * from './NewProposal'
+export * from './NewProposalTitleDescriptionHeader'
 export * from './PaginatedProposalVotes'
 export * from './ProgressBar'
 export * from './ProposalCard'
